fix(scrawler mail): fix can't find vericode email after change inbox folder
Because I can't bear lots of vericode email from school, I added a rule to move vericode emails to a
dedicated folder. However, this caused the scrawler can't locate vericode emails since they are not
in default INBOX folder. So, select the inbox before searching emails would be the fix.